+++ components/Content/Gallery.tsx
@@ -2,7 +2,6 @@ import React, { useEffect, useState } from "react";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { config } from "@fortawesome/fontawesome-svg-core";
 import "@fortawesome/fontawesome-svg-core/styles.css";
-import { iconMapper, getType, getIcon, getName } from "../../helpers/gallery.helpers";
 import Pagination from "../Dashboard/Pagination";
 config.autoAddCss = false;
 import { SubmarinedContent } from "../../types/SubmarinedContent";
@@ -10,17 +9,49 @@ import { useAppDispatch } from "../../store/hooks";
 import { setSubmarinedContent } from "../../store/slices/submarinedContentSlice";
 import { getKy } from "../../helpers/ky";
 import SingleMediaDisplay from "./SingleMediaDisplay";
+import { faImage, faMusic, faVideo } from "@fortawesome/free-solid-svg-icons";
+import mime from "mime";
 
 interface GalleryProps {
   content: SubmarinedContent;
   name: string;
 }
+export const iconMapper = (type) => {
+  const map = {
+    image: faImage,
+    audio: faMusic,
+    video: faVideo,
+    "application/pdf": "fa-file-pdf-o",
+    "application/msword": "fa-file-word-o",
+    "application/vnd.ms-word": "fa-file-word-o",
+    "application/vnd.oasis.opendocument.text": "fa-file-word-o",
+    "application/vnd.openxmlformats-officedocument.wordprocessingml": "fa-file-word-o",
+    "application/vnd.ms-excel": "fa-file-excel-o",
+    "application/vnd.openxmlformats-officedocument.spreadsheetml": "fa-file-excel-o",
+    "application/vnd.oasis.opendocument.spreadsheet": "fa-file-excel-o",
+    "application/vnd.ms-powerpoint": "fa-file-powerpoint-o",
+    "application/vnd.openxmlformats-officedocument.presentationml": "fa-file-powerpoint-o",
+    "application/vnd.oasis.opendocument.presentation": "fa-file-powerpoint-o",
+    "text/plain": "fa-file-text-o",
+    "text/html": "fa-file-code-o",
+    "application/json": "fa-file-code-o",
+    "application/gzip": "fa-file-archive-o",
+    "application/zip": "fa-file-archive-o",
+  };
+  return map[type];
+};
+
+export const getType = (type) => {
+  return mime.getType(type);
+};
+
 
 export default function Gallery({ content, name }: GalleryProps) {
   const [items, setItems] = useState([]);
   const [offset, setOffset] = useState(0);
   const [isDisplaying, setIsDisplaying] = useState<boolean>(false)
   const [displayItem, setDisplayItem] = useState(null)
+  const mainThree = ["image", "audio", "video"];
   const limit = 10;
   const dispatch = useAppDispatch();
 
@@ -68,6 +99,22 @@ export default function Gallery({ content, name }: GalleryProps) {
     setIsDisplaying(true)
     setDisplayItem(item)
   }
+
+  const getIcon = (filename) => {
+    const extension = filename.substr(filename.lastIndexOf(".") + 1);
+    const type = getType(extension);
+    let icon = null;
+    if (type && mainThree.includes(type.split("/")[0])) {
+      icon = iconMapper(type.split("/")[0]);
+    } else {
+      icon = iconMapper(type);
+    }
+    return icon;
+  };
+  const getName = (name) => {
+    const items = name.split("/");
+    return items[items.length - 1];
+  };
 
   return (
     <div className="bg-white">
